+++ public/plugins/select2-develop/src/js/select2/selection/single.js
@@ -35,7 +35,7 @@ define([
     this.$selection.find('.select2-selection__rendered')
       .attr('id', id)
       .attr('role', 'textbox')
-      .attr('aria-readonly', 'true');
+      .attr('aria-', 'true');
     this.$selection.attr('aria-labelledby', id);
 
     this.$selection.on('mousedown', function (evt) {