+++ resources/views/ddad/shops/index.blade.php
@@ -83,6 +83,7 @@
                                         <div class="st_table_action_btn_wrap">
                                             <button class="st_table_action_btn dropdown-toggle" data-toggle="dropdown"><i class="material-icons">more_horiz</i></button>
                                             <div class="dropdown-menu dropdown-size-sm dropdown-menu-right st_boxshadow">
+                                                <a class="dropdown-item" href="{{ route('shops.show', $shop) }}"><i class="material-icons-outlined">visibility</i>View</a>
                                                 <a class="dropdown-item" href="{{ route('shops.edit', $shop) }}"><i class="material-icons-outlined">create</i>Edit</a>
                                                 <a class="dropdown-item" href="" onclick="" data-delete_action="{{ route('shops.destroy', $shop) }}"><i class="material-icons-outlined">delete_outline</i>Delete</a>
                                             </div>